+++ packages/suspense-website/src/examples/createStreamingCache/hook.ts
@@ -6,7 +6,7 @@ function ReactComponent({ path }: { path: string }) {
   const stream = exampleStreamingCache.stream(path);
 
   // Uses useSyncExternalStore to subscribe to changes
-  const { complete, progress, values } = useStreamingValues(stream);
+  const { complete, progress, status, values } = useStreamingValues(stream);
 
   // ...
 }