+++ packages/suspense/src/hooks/useCacheStatus.ts
@@ -5,7 +5,7 @@ import { Cache, Status } from "../types";
 export function useCacheStatus<Params extends Array<any>>(
   cache: Cache<Params, any>,
   ...params: Params
-): Status | undefined {
+): Status {
   return useSyncExternalStore<Status | undefined>(
     (callback) => cache.subscribeToStatus(callback, ...params),
     () => cache.getStatus(...params),